Optimize Loading Mechanism
