+++ src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,3 @@
 de.viadee.bpm.externaltask.retry.aspect.config.ExternalTaskRetryAspectConfiguration
+de.viadee.bpm.externaltask.retry.aspect.config.CamundaExternalTaskRetryAspectConfiguration
+de.viadee.bpm.externaltask.retry.aspect.config.OperatonExternalTaskRetryAspectConfiguration
