+++ src/test/java/de/viadee/bpm/operaton/externaltask/retry/aspect/behaviour/InvalidDefaultBehaviourTestOperaton.java
@@ -35,15 +35,29 @@ import de.viadee.bpm.operaton.externaltask.retry.aspect.OperatonBaseTest;
 import org.junit.jupiter.api.Test;
 import org.springframework.test.context.TestPropertySource;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.Mockito.when;
 
+@TestPropertySource(properties = "de.viadee.bpm.operaton.external-task.retry-config.default-behavior=R3/PT1D")
+public class InvalidDefaultBehaviourTestOperaton extends OperatonBaseTest {
 
-@TestPropertySource(properties = "de.viadee.bpm.operaton.external-task.retry-config.identifier=CUSTOM_SOMETHING")
-public class CustomRetryTimeCycleIdentifierTestCamunda extends OperatonBaseTest {
 
     @Test
-    public void customRetryTimeCycleIdentifier() {
-        assertEquals("CUSTOM_SOMETHING", this.properties.getIdentifier());
+    public void runtimeException() {
+        // prepare
+        when(this.externalTask.getRetries()).thenReturn(null);
+        when(this.externalTask.getExtensionProperty(this.properties.getIdentifier())).thenReturn("invld!");
+
+        // test
+        this.operatonExternalTaskRetryAspect.handleErrorAfterThrown(this.joinPoint, new RuntimeException(), this.externalTask, this.externalTaskService);
+
+        // verify
+        this.verifyNoBpmnErrorAtAll();
+        this.verifyHandleFailure();
+
+        // assert
+        this.assertRemainingRetries(3);
+        this.assertNextRetryInterval(5 * MINUTES_TO_MILLIS);
     }
+
 
 }
